Updated the background images of register page and imlemented search api in the leave page
