+++ hrms-frontend/src/constants/commonFunction.ts
@@ -4,3 +4,13 @@ export const getLabelByValue = (value: string, employeProfetion: any) => {
   );
   return profession ? profession.label : "Value not found";
 };
+
+export const DateFormatConvertor = (date: any) => {
+  const newDate = new Date(date)
+
+  const formateddate = newDate.getDate();
+  const cformateddate = formateddate >= 10 ? formateddate : `0${formateddate}`
+  const formatedMonth = newDate.getMonth() + 1;
+  const formatedYear = newDate.getFullYear();
+  return `${formatedYear}/${formatedMonth}/${cformateddate}`
+}
