+++ hrms-frontend/src/redux/leave/leaveSelector.ts
@@ -1,0 +1,3 @@
+export const manageLeaveSelector = (state: any) => {
+    return state.leaveSlice;
+};
