+++ hrms-frontend/src/redux/user/userSelector.ts
@@ -1,0 +1,3 @@
+export const manageUserSelector = (state: any) => {
+  return state.userSlice;
+};
