--- hrms-frontend/src/services/user/slices/allLeaves/leaveSelector.ts
@@ -1,3 +1,0 @@
-export const manageLeaveSelector = (state: any) => {
-    return state.leaveSlice;
-};
