--- hrms-frontend/src/services/user/slices/allUser/userSelector.ts
@@ -1,3 +1,0 @@
-export const manageUserSelector = (state: any) => {
-  return state.userSlice;
-};
